scar/requirements.txt
# Basic scientific computing
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Table output beautification
tabulate>=0.8.9

# Single-cell data structure support
anndata>=0.8.0
h5py>=3.6.0
#pyreadr>=0.4.0    # For reading .rds/.RData files (requires C++ compilation environment)

# Deep learning
torch>=1.12.0     # Determine version based on your current CUDA environment

# Machine learning and image processing (if needed)
scikit-learn>=1.0.0
scikit-image>=0.19.0

loompy
nvidia-ml-py3 #pynvml package
pyyaml


#pip install -r requirements.txt
scanpy